1105 fix project level report (#1267)
* Move checkValueSatisfiesCondition to @tupaia/utils

* Fix tests

* Use new config for old databuilders

* Fix project level report unfpa
