+++ packages/web-config-server/src/apiV1/dataBuilders/helpers/checkAgainstConditions.js
@@ -1,40 +1,5 @@
 import isPlainObject from 'lodash.isplainobject';
-
-const OPERATOR_TO_VALUE_CHECK = {
-  '=': (value, target) => value === target,
-  '>=': (value, target) => value >= target,
-  '<=': (value, target) => value <= target,
-  '>': (value, target) => value > target,
-  '<': (value, target) => value < target,
-  range: (value, target) => target[0] <= value && value <= target[1],
-  rangeExclusive: (value, target) => target[0] < value && value < target[1],
-  regex: (value, target) => !!value.match(target),
-  in: (value, target) => target.includes(value),
-};
-
-const NUMERIC_OPERATORS = ['>=', '<=', '>', '<', 'range', 'rangeExclusive'];
-
-const ANY_VALUE_CONDITION = '*';
-
-export const checkValueSatisfiesCondition = (value, condition) => {
-  if (value === undefined) return false;
-
-  if (!isPlainObject(condition)) {
-    return (condition === ANY_VALUE_CONDITION && value !== '') || value === condition;
-  }
-
-  const { operator, value: targetValue } = condition;
-
-  const checkValue = OPERATOR_TO_VALUE_CHECK[operator];
-  if (!checkValue) {
-    throw new Error(`Unknown operator: '${operator}'`);
-  }
-
-  // Prevents '' (no data) being cast to 0 and giving incorrect results
-  if (NUMERIC_OPERATORS.includes(operator) && value === '') return false;
-
-  return checkValue(value, targetValue);
-};
+import { checkValueSatisfiesCondition } from '@tupaia/utils';
 
 /**
  * @param {Event[]} events
